fix: use new linking status in component to track link creation status
When a link is created, i.e. when it is determined that the injected information
is not already in the target deployment, we update the target component's status
with the link information so that we can track that a link is being created and
put the component in the reconcile loop until the new pod is created.
Conversely, when the new pod is created, we update the component's linking status
and notify the incoming link that the process is done so that it can be
reconciled as well.