fix: default parameters are not working, this is a temporary fix
The issue here is that we need to somehow pass information from the
component that's requesting the capability to the capability itself.
That information, however, is only pertinent for this particular binding
so it is wrong to add the parameters to the capability itself. We need
to find a better way to do the actual binding and transfer the
information: maybe using an internal CR to record the binding
information or maybe a Service Binding request?